Add PARTIALS --partials flag, correct phase adjustment, write as floats
If PARTIALS is defined at compilation tine, you get a --partials flag
which writes the result of every iFFT to a separate .wav file.

The hack to get successive frames' frequency elements in phase needed
depends on the actual frequency represented by the bin of choice, not
the frequency it should be.  This corrects phasing of adjacent frames.

As we are getting overflows, we write the WAV file as floats, not shorts,
so we can reduce the volume in a wave editor and see better what is going on.