Drop phase correction as it causes worse artifacts than it cures
We'd like to ensure that waves for the same bin in successive frames are
in phase with each other, but this makes the partials overflow, presumably
because all the component sin or cos waves coincide in phase so sum up.
Haven't found a way to avoid this yet, except by dropping the phase correction
completely.