Improve dB value for missing colours by interpolating where between.
When there are colors in the graph that are nor èresent in the scale,
this improves the pick-the-nearest-color fallthrough case by spotting
when a hue is between two elements of the color scale, and interpolating
between them.
Also, values off the bottom of the scale are handled with interpolation
where possible.