Add --fill flag to interpolate FFT inputs instead of sparsing them
The FFT input array usually has more points per Hz than the pixels in one
column, and usually we take each color point and set one point in the FFT
input array, that of the nearest frequency value, giving a sparse FFT input.
This flag makes it fill every entry in the FFT input array by interpolating
between the existent point values.

Without --fill, you tend to get a flutter at the same frequency as the
the number of input columns per second because the frequency bands derived
from the pixel rows are the same distance apart as the number of pixel columns
per second, creating beat frequencies between adjacent synthesized tones.

With --fill you tend to get a flutter at a frequency that is the distance
between FFT input buckets, which are more finely spaced in frequency than
the pixel rows, so the flutter due to beats between adjacent frequencies is
slower. I'm not sure which is worse.